Fixing type error

+++ htrc/util/__init__.py
@@ -12,7 +12,7 @@ def split_items(seq, split_size):
     :param seq: A sequence to split.
     :param split_size: The maximum size of each split.
     """
-    full_segments = math.floor(len(seq) / split_size)
+    full_segments = int(math.floor(len(seq) / split_size))
     for i in range(1,full_segments+1):
         yield seq[(i-1)*split_size:i*split_size]
     if (full_segments * split_size) < len(seq):
